+++ src/commands/Minion/daily.ts
@@ -21,7 +21,7 @@ const { triviaQuestions } = JSON.parse(
 
 import { BotCommand } from '../../lib/BotCommand';
 import { COINS_ID, Emoji, SupportServer, Time } from '../../lib/constants';
-import pets from '../../lib/pets';
+import pets from '../../lib/data/pets';
 import { ClientSettings } from '../../lib/settings/types/ClientSettings';
 import { UserSettings } from '../../lib/settings/types/UserSettings';
 import dailyRoll from '../../lib/simulation/dailyTable';
