+++ src/commands/OSRS_Fun/pet.ts
@@ -2,7 +2,7 @@ import { CommandStore, KlasaMessage } from 'klasa';
 import { cleanString } from 'oldschooljs/dist/util';
 
 import { BotCommand } from '../../lib/BotCommand';
-import pets from '../../lib/pets';
+import pets from '../../lib/data/pets';
 import { roll } from '../../lib/util';
 
 export default class extends BotCommand {
