+++ src/monitors/petmessages.ts
@@ -1,6 +1,6 @@
 import { KlasaMessage, Monitor, MonitorStore } from 'klasa';
 
-import pets from '../lib/pets';
+import pets from '../lib/data/pets';
 import { GuildSettings } from '../lib/settings/types/GuildSettings';
 import { UserSettings } from '../lib/settings/types/UserSettings';
 import { roll } from '../lib/util';
